ovn: Don't override idl lookup() method with timeout
The timeout and notify_handler parameters are removed in ovsdbapp [1]
because they have never really been used and if we tried to use it, it
caused lots of various issues bringing low benefit. We often call
lookup() on non-existing objects and adding 2 seconds wait interval has
terrible effect on performance.

[1] https://review.opendev.org/c/openstack/ovsdbapp/+/835151

Change-Id: I5f22ecfe723e8710ff62e7e1a006e534ccc93919
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>