Make the recheck policy more prominently visible
It has been suggested in the Neutron CI meeting to include a section in
the documentation advicing against blind rechecks. It turns out that
such section already exists. What this change does is to move the
section to the first level of the contributors guide, to make it more
visible. This change also improves some wording and adds some examples
of proper recheck comments.

Change-Id: Ib0a00d13a28f98b0a0f26c7233365d04453db4e0