doc: point to relevant nova config in routed networks guide
While the focus of the routed provider networks documentation
is on the Neutron side of things, this document provides a useful
location to point at relevant Nova configuration.

This reference relates to the Nova scheduler and ensuring that
instance move operations don't result in instances being scheduled
to hosts which do not have access to the required network(s).

Change-Id: Idd7382bd38d70383107d2997205ff880bede8d96