Docs build: Add requirements.txt to deps
Since we use usedevelop=True, as a result neutron
installation happens at 'develop-inst' phase and it installs
requirements from requirements.txt without considering
tox constraints, This can lead to issue where broken
requirements are installed in docs job and as a result
it will fail.

In order to get constraints be considered for this case, add
requirements.txt to deps for docs env.

Change-Id: Ic8f6472fe708d12f8fc129729465367083e1095a