Avoid register config options on imports
Continue similar approach following in [1], where some project imports
collide with config options.

As part of the change, a wrapped decorator has been implemented to cover
those functions that include any of the ovn config options as value to
the decorators arguments (e.g. tenacity retry). This way we avoid
requiring the options to be registered as soon as the module is imported,
where they have not yet been registered by a main process.

[1] https://review.opendev.org/c/openstack/neutron/+/837392

Co-authored-by: Jakub Libosvar <libosvar@redhat.com>
Co-authored-by: Fernando Royo <froyo@redhat.com>

Change-Id: I4bccb094ee7f690cbc352c38b5b39d505e6ea460