ovn: Change logging level for OVN metadata agent
The provisioning message is in DEBUG level while unprovisioning in INFO.
When agent is running with INFO level, it produces messages about ports
being bound, unbound and when cleaning up namespace but no messages
about creating the namespaces.

This patch makes sure that on INFO level we get information about both
provisioning and unprovisioning datapaths. These messages appear only
with first or last ports respectively on the node and therefore are not
chatty.

Change-Id: I1851587e49822465e1308390a9810ee21535245f