Skip "PortBindingChassisEvent" if revision number changes
Since [1], the "external_ids" of the NB Logical_Router_Port register
are copied into the SB Port_Binding "external_ids". When a change
in a Port_Binding register is received, if only the
"external_ids:revision_number" is changed, we skip any update on the
related Logical_Router_Port.

If not, that will lead to an infinite loop: Neutron will update
the Logical_Router_Port with a new revision number and OVN will
copy this new revision number to the SB register, triggering again
the update of the NB Logical_Router_Port

[1]https://www.mail-archive.com/ovs-dev@openvswitch.org/msg62836.html

Closes-Bug: #1973347
Change-Id: Ib51764778a666050c42de0dfeb9bf9b185d44bb7