ovn-migration: Remove second tripleo-update call
The second tripleo-update call was there only to update back integration
bridge variable. However the action did the whole deploy for all nodes
in the environment, which can take hours to finish. This patch relies on
other patch in THT that sets br-int as default value. This way once
ovn-extras.yml is not used and the integration bridge is not overriden,
we no longer need to update tripleo about the value so it's safe to
remove the second call.

This patch is backported only to Wallaby stable branch (no Xena or Yoga) as it depends on the TripleO patch linked below.

Depends-On: https://review.opendev.org/c/openstack/tripleo-heat-templates/+/847999
Change-Id: I7455852898bb6e9698ecf5261d4401bc5077c506
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>
(cherry picked from commit 60fade43de9555aab110d898ca27684bcaa975be)