[Trunk] Check if parent port belongs to OVN to activate
In Trunk service, the OVN was setting always the trunk to "ACTIVE"
anytime the trunk was created or updated. This patch changes this
behaviour: only when the trunk parent port belongs to OVN (that
means, there is a related logical switch port to this Neutron port),
the methods will set the trunk to "ACTIVE".

Closes-Bug: #1974183
Change-Id: Ied5ef53882d4249e0ceaa731239befdc3ba67d03
(cherry picked from commit 68bbd3586bdc313a6b67f2983ba2331a7956856c)