[sqlalchemy-20] Use session.transaction information to decide if active
The "session.transaction" member "_connections" is a dictionary with the
active database connections of the current transaction. Neutron will use
it to determine if the session is active or not. This parameter will
provide more accurate information than "new", "dirty" or "deleted". A
read (SELECT) operation won't be included on those lists while the
transaction is currently active.

Closes-Bug: #1975542
Change-Id: Ibb75b324950c639d30c77d4e5269a8c53d6a0bb5