Revert "Use Port_Binding up column to set Neutron port status"
This reverts commit 37d4195b516f12b683b774f0561561b172dd15c6.
Conflicts:
 	neutron/common/ovn/constants.py
 	neutron/plugins/ml2/drivers/ovn/mech_driver/ovsdb/ovsdb_monitor.py

Also revert below 2 commits which were added on
top of the parent commit:-

Revert "Ensure subports transition to DOWN"
This reverts commit 5e036a6b281e4331f396473e299b26b2537d5322.

Revert "Ensure only the right events are processed"
This reverts commit 553f462656c2b7ee1e9be6b1e4e7c446c12cc9aa.

Reason for revert: These patches have caused couple of issues[1][2][3].
[1][2] are same issue just one is seen in c8/c9-stream and other in
rhel8 and both contains much info about the issue.
[3] is currently happening only in rhel8/rhel9 as this issue is visible
only with the patch in revert and ovn-2021>=21.12.0-55(fix of [4]) which
is not yet available in c8/c9-stream.

[1][2] happens randomly as the patch under revert has moved the
events to SB DB which made a known OVN issue[5] occur more often as in
that issue SB DB Event queue floods with too many events of
PortBindingChassisEvent making other events like PortBindingUpdateUpEvent
to wait much longer and hence triggering VirtualInterfaceCreateException.

NB DB Event queue is different and hence with revert we are trying to
lower the side effect of the OVN issue[5].

This patch can be re reverted once [3] and [5] are fixed.

[1] https://bugs.launchpad.net/tripleo/+bug/1964940/
[2] https://bugzilla.redhat.com/show_bug.cgi?id=2081631
[3] https://bugzilla.redhat.com/show_bug.cgi?id=2090604
[4] https://bugzilla.redhat.com/show_bug.cgi?id=2037433
[5] https://bugzilla.redhat.com/show_bug.cgi?id=1974898

Closes-Bug: #1964940
Closes-Bug: rhbz#2081631
Closes-Bug: rhbz#2090604
Related-Bug: rhbz#2037433
Related-Bug: rhbz#1974898
Change-Id: I159460be27f2c5f105be4b2865ef84aeb9a00094