[OVS][QoS] Unset the min-bw QoS from the phys bridge interface
The minimum bandwidth OVS QoS policy is set in the physical bridge
interface, not the port that receives the Neutron QoS policy.
If no other port has a minimum bandwidth rule, the OVS QoS policy
can be unset from this port and deleted. This patch fixes the
port that must be updated.

Closes-Bug: #1977819
Change-Id: I0ebcc85f5ffa9dc29ffedee0c7df1828c85e1576