[FT] Ensure "update_virtual_port_host" has been called
An OVS DB event once matched, is stored in a cache and processed in
``ovsdbapp.event.RowEventHandler.notify_loop``. The method
``update_virtual_port_host`` is called in the ``run`` method of the
event. Sometimes in the FT execution, the test checks if the
method has been called when the event ``run`` method has not been
executed yet.

This patch adds a short active wait to ensure the event is actually
processed and the method called.

This bug can be easily reproduced if in the
``PortBindingUpdateVirtualPortsEvent.run`` method, a 2 seconds delay
is added:
  import time; time.sleep(2)

Closes-Bug: #1971672
Change-Id: I2e389e3cce70505ce01f03c09b36fa9d55be964a