Use common wait_until_ha_router_has_state method everywhere
In the L3 functional tests framework module there is already helper
method called wait_until_ha_router_has_state which should be used to
wait for desired HA router's state.
This method has proper debug logging added so debugging issues in CI is
easier when it's used.
It is also decorated with unstable_test decorator to skip tests when
router will fail to transition to desired state (see related bug for
details).

In some tests this method wasn't used so we couldn't benefit from the
logging and unstable_test decorator there. Now it should be unifed and
used everywhere in the same way.

Related-Bug: #1956958
Change-Id: I9d79b123bb20ded327208d84a14d4f8d2e505087