QoS minimum bandwidth documentation: OVN backend
Updated documentation about minimum guaranteed bandwidth support
for OVN backend. This mech driver is capable of informing the
Placement API about the bandwidth resources of each chassis and
schedule a port depending on the minimum bandwidth rules applied.

This patch also updates the document pointing to the new
``resource_requeste`` blob passed to Nova with the information of
the requested resources of each port.

Related-Bug: #1578989
Change-Id: I75401fa64e1ccd814aa485a74b38092849ff2a18