update neutron-linuxbridge-tempest for bug 1980948
change Ib18efa3f472736b58c8967847b1061da0e3897d7
made linuxbridge an experimental feature but
it did not enable linuxbridge in the linuxbridge jobs

this job is used by both nova and devstack and perhaps
other project on an as needed basis. As a result
this is causing gate failures in neutron when some files
are modifed such as the netwroking module or vif.py files

This change enabled linux brdige in the base linux bridge
job and removes an old duplicate job defintion since
nova and devstack now use the new name.

Change-Id: I290027fbac47bd1df7bcdd08c405db107ed63db0
Closes-Bug: #1980948