Test: mock out _check_netfilter_for_bridges in unit tests
_check_netfilter_for_bridges do some calls which are not for unit tests.
Specifically the call for sysctl -N net.bridge seems to be broken with
a recent kernel upgrade on bionic (see the irc log for this discussion
in [1]).
Add OS_TEST_TIMEOUT:180 to tox testenv, this can help to debug such
issues.
During the rootwrap to privsep change some of these mocks were added,
(see [2]) so backporting (before Wallaby mostly) can be done with
caution.

[1]: https://meetings.opendev.org/irclogs/%23openstack-infra/%23openstack-infra.2022-07-20.log.html#t2022-07-20T12:59:32
[2]: https://review.opendev.org/c/openstack/neutron/+/764015/12/neutron/tests/unit/agent/linux/openvswitch_firewall/test_iptables.py

Change-Id: Ic366d67770cc6e19a4412eaf2f1ab53780e98ee8
Closes-Bug: #1982206