[stable/train] Convert neutron-grenade job to the zuul v3 job
Other grenade jobs, like neutron-grenade-multinode are already converted
to the native zuul v3 definitions but neutron-grenade job wasn't and it
seems that as it is using devstack-gate project to provision vm, for
some reason it stucks now and job fails with timeout.
To avoid that lets convert that job to the native Zuul v3 syntax.

This patch is needed only for stable/train as in newer branches we don't
have that single node job anymore.

This patch also switches neutron-grenade job to be voting in the check queue
and to be run in the gate queue. It was made non-voting and no gating with
the patch https://review.opendev.org/c/openstack/neutron/+/851330 to
workaround the issue which this patch fixes.

Closes-bug: #1982720
Change-Id: Iaa42e0d3020c67d22583ed70e624c401b192dc93